requirements.txt
anyio==4.7.0
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot[job-queue]==21.9
sniffio==1.3.1
beautifulsoup4==4.12.3
APScheduler>=3.6.3
python-dotenv==1.0.1
